Move a global JSON networks config. Merge wrapped currencies with it
