+++ src/networks.json
@@ -1,0 +1,80 @@
+{
+  "1": {
+    "name": "Ethereum",
+    "rpc": "https://mainnet.infura.io/v3/212957e8adf14bf8aecf82358083e63e",
+    "chainId": 1,
+    "wrapperCurrency": "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
+  },
+  "3": {
+    "name": "Ropsten",
+    "rpc": "https://ropsten.infura.io/v3/212957e8adf14bf8aecf82358083e63e",
+    "chainId": 3,
+    "wrapperCurrency": "0xc778417e063141139fce010982780140aa0cd5ab"
+  },
+  "4": {
+    "name": "Rinkeby",
+    "rpc": "https://rinkeby.infura.io/v3/212957e8adf14bf8aecf82358083e63e",
+    "chainId": 4,
+    "wrapperCurrency": "0xc778417e063141139fce010982780140aa0cd5ab"
+  },
+  "56": {
+    "name": "Binance Smart Chain",
+    "rpc": "https://bsc-dataseed1.ninicoin.io",
+    "chainId": 56,
+    "wrapperCurrency": "0xbb4cdb9cbd36b01bd1cbaebf2de08d9173bc095c"
+  },
+  "97": {
+    "name": "BSC Testnet",
+    "rpc": "https://data-seed-prebsc-1-s1.binance.org:8545/",
+    "chainId": 97,
+    "wrapperCurrency": "0xae13d989dac2f0debff460ac112a837c89baa7cd"
+  },
+  "137": {
+    "name": "Polygon",
+    "rpc": "https://rpc-mainnet.matic.network",
+    "chainId": 137,
+    "wrapperCurrency": "0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270"
+  },
+  "80001": {
+    "name": "Polygon testnet",
+    "rpc": "https://rpc-mumbai.maticvigil.com",
+    "chainId": 80001,
+    "wrapperCurrency": "0x9c3c9283d3e44854697cd22d3faa240cfb032889"
+  },
+  "42161": {
+    "name": "Arbitrum",
+    "rpc": "https://arb1.arbitrum.io/rpc",
+    "chainId": 42161,
+    "wrapperCurrency": ""
+  },
+  "421611": {
+    "name": "Arbitrum testnet",
+    "rpc": "https://rinkeby.arbitrum.io/rpc",
+    "chainId": 421611,
+    "wrapperCurrency": ""
+  },
+  "88": {
+    "name": "TomoChain",
+    "rpc": "https://rpc.tomochain.com",
+    "chainId": 88,
+    "wrapperCurrency": ""
+  },
+  "30": {
+    "name": "RSK",
+    "rpc": "https://public-node.rsk.co",
+    "chainId": 30,
+    "wrapperCurrency": ""
+  },
+  "31": {
+    "name": "RSK Testnet",
+    "rpc": "https://public-node.testnet.rsk.co",
+    "chainId": 31,
+    "wrapperCurrency": ""
+  },
+  "100": {
+    "name": "xDAI Chain",
+    "rpc": "https://rpc.xdaichain.com",
+    "chainId": 100,
+    "wrapperCurrency": ""
+  }
+}
